Adds first-pass fluent-like API for command building
The current implementation is rather difficult to reason about, much of it is not extendable, and is particularly difficult to change or debug.

This addition is meant to replace the old builder once complete, and will be created i several steps. It aims to provide a proper builder API that is easy to extend and change. It's more verbose, which is part of the goal of ensuring that the call flow is more linear.