Moves several models to /Models and their interfaces to /Models/Interfaces
The decision is driven by the need to have models & their interfaces organized in a logical manner where it it easy to find one or the other.

For a small project models residing in the same namespace makes them much easier to reason about than if they are scattered in other namespaces/directories.