Splits command builders into their own files. Prefixes with FFmpeg
The prefixing with `FFmpeg` is for clarity, though it does come at the cost of verbose type names.

It is a questionable naming convention, but it does meet the goal of being semantic.

The interfaces where originally created with the thought that the builders would need to be more complex than they are. They are simpler than planned, which means the interfaces are actually unnecessary for the creation of a fluent api.