Fixes several bugs with input detection
Splits supported input an output codecs, adds h265 and theora to supported input codecs for detection.
Always applies a default dash key interval of 3s if none is given.
Avoids detecting the first and last dot component (base filename and extension) as a language in auxiliary subtitle files.
Tries harder to get a framerate and bitrate from the input.
Converts string framerates with slash symbols like 25/1 to decimal instead of defaulting to zero.